+++ source/pid_iv_tid.c
@@ -103,7 +103,7 @@ IWRAM_CODE MAX_OPTIMIZE u32 generate_ot(u16 tid, u8* name) {
     u16 sid = 0;
     seed = get_next_seed(seed);
     
-    for(int i = 0; i < OT_NAME_GEN3_SIZE; i++) {
+    for(size_t i = 0; i < OT_NAME_GEN3_SIZE; i++) {
         if(name[i] == GEN3_EOL)
             break;
         seed += name[i];
