+++ source/text_handler.c
@@ -2,6 +2,7 @@
 #include "text_handler.h"
 #include "bin_table_handler.h"
 #include "useful_qualifiers.h"
+#include <stddef.h>
 
 #include "trainer_names_bin.h"
 #include "text_gen3_to_gen12_int_bin.h"
@@ -20,18 +21,18 @@
 #define GENERIC_TO_UPPER 0x20
 #define GEN12_TRAINER 0x5D
 
-u8 text_general_count_question(const u8*, u8, u8, u8);
-u8 text_general_size(const u8*, u8, u8);
-void text_general_conversion(const u8*, u8*, u8, u8, u8, u8, const u8*);
-u8 text_general_is_same(const u8*, const u8*, u8, u8, u8);
-void text_general_copy(const u8*, u8*, u8, u8, u8);
-void text_general_concat(const u8*, const u8*, u8*, u8, u8, u8, u8);
-void text_general_replace(u8*, u8, u8, u8, u8);
-void text_general_terminator_fill(u8*, u8, u8);
+size_t text_general_count_question(const u8*, size_t, u8, u8);
+size_t text_general_size(const u8*, size_t, u8);
+void text_general_conversion(const u8*, u8*, size_t, size_t, u8, u8, const u8*);
+u8 text_general_is_same(const u8*, const u8*, size_t, size_t, u8);
+void text_general_copy(const u8*, u8*, size_t, size_t, u8);
+void text_general_concat(const u8*, const u8*, u8*, size_t, size_t, size_t, u8);
+void text_general_replace(u8*, size_t, u8, u8, u8);
+void text_general_terminator_fill(u8*, size_t, u8);
 
-u8 text_general_count_question(const u8* src, u8 src_size, u8 terminator, u8 question) {
-    int counter = 0;
-    for(int i = 0; i < src_size; i++) {
+size_t text_general_count_question(const u8* src, size_t src_size, u8 terminator, u8 question) {
+    size_t counter = 0;
+    for(size_t i = 0; i < src_size; i++) {
         if(src[i] == terminator)
             break;
         if(src[i] == question)
@@ -40,16 +41,16 @@ u8 text_general_count_question(const u8* src, u8 src_size, u8 terminator, u8 que
     return counter;
 }
 
-u8 text_general_size(const u8* src, u8 src_size, u8 terminator) {
-    int i = 0;
+size_t text_general_size(const u8* src, size_t src_size, u8 terminator) {
+    size_t i = 0;
     for(; i < src_size; i++)
         if(src[i] == terminator)
             break;
     return i;
 }
 
-void text_general_conversion(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 src_terminator, u8 dst_terminator, const u8* conv_table) {
-    for(int i = 0; (i < src_size) && (i < dst_size); i++) {
+void text_general_conversion(const u8* src, u8* dst, size_t src_size, size_t dst_size, u8 src_terminator, u8 dst_terminator, const u8* conv_table) {
+    for(size_t i = 0; (i < src_size) && (i < dst_size); i++) {
         if(src[i] == src_terminator) {
             dst[i] = dst_terminator;
             break;
@@ -60,11 +61,11 @@ void text_general_conversion(const u8* src, u8* dst, u8 src_size, u8 dst_size, u
         dst[src_size] = dst_terminator;
 }
 
-u8 text_general_is_same(const u8* src, const u8* dst, u8 src_size, u8 dst_size, u8 terminator) {
+u8 text_general_is_same(const u8* src, const u8* dst, size_t src_size, size_t dst_size, u8 terminator) {
     if(text_general_size(src, src_size, terminator) != text_general_size(dst, dst_size, terminator))
         return 0;
 
-    int i = 0;
+    size_t i = 0;
     for(; (i < src_size) && (i < dst_size); i++) {
         if(src[i] == terminator) {
             if(dst[i] == terminator)
@@ -81,8 +82,8 @@ u8 text_general_is_same(const u8* src, const u8* dst, u8 src_size, u8 dst_size, 
     
 }
 
-void text_general_copy(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 terminator) {
-    for(int i = 0; (i < src_size) && (i < dst_size); i++) {
+void text_general_copy(const u8* src, u8* dst, size_t src_size, size_t dst_size, u8 terminator) {
+    for(size_t i = 0; (i < src_size) && (i < dst_size); i++) {
         if(src[i] == terminator) {
             dst[i] = terminator;
             break;
@@ -93,15 +94,15 @@ void text_general_copy(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 term
         dst[src_size] = terminator;
 }
 
-void text_general_concat(const u8* src, const u8* src2, u8* dst, u8 src_size, u8 src2_size, u8 dst_size, u8 terminator) {
-    int i = 0;
+void text_general_concat(const u8* src, const u8* src2, u8* dst, size_t src_size, size_t src2_size, size_t dst_size, u8 terminator) {
+    size_t i = 0;
     for(; (i < src_size) && (i < dst_size); i++) {
         if(src[i] == terminator) {
             break;
         }
         dst[i] = src[i];
     }
-    for(int j = 0; (j < src2_size) && ((i+j) < dst_size); j++) {
+    for(size_t j = 0; (j < src2_size) && ((i+j) < dst_size); j++) {
         if(src2[j] == terminator) {
             dst[i+j] = terminator;
             break;
@@ -112,9 +113,9 @@ void text_general_concat(const u8* src, const u8* src2, u8* dst, u8 src_size, u8
         dst[i+src2_size] = terminator;
 }
 
-void text_general_replace(u8* src, u8 src_size, u8 base_char, u8 new_char, u8 terminator) {
+void text_general_replace(u8* src, size_t src_size, u8 base_char, u8 new_char, u8 terminator) {
     u8 replaced_once = 0;
-    for(int i = 0; i < src_size; i++) {
+    for(size_t i = 0; i < src_size; i++) {
         if((src[i] == terminator) && (base_char != terminator))
             break;
         if(src[i] == base_char) {
@@ -126,110 +127,110 @@ void text_general_replace(u8* src, u8 src_size, u8 base_char, u8 new_char, u8 te
     }
 }
 
-void text_general_terminator_fill(u8* src, u8 src_size, u8 terminator) {
-    for(int i = 0; i < src_size; i++)
+void text_general_terminator_fill(u8* src, size_t src_size, u8 terminator) {
+    for(size_t i = 0; i < src_size; i++)
         src[i] = terminator;    
 }
 
-void text_generic_concat(const u8* src, const u8* src2, u8* dst, u8 src_size, u8 src2_size, u8 dst_size) {
+void text_generic_concat(const u8* src, const u8* src2, u8* dst, size_t src_size, size_t src2_size, size_t dst_size) {
     text_general_concat(src, src2, dst, src_size, src2_size, dst_size, GENERIC_EOL);
 }
 
-void text_gen3_concat(const u8* src, const u8* src2, u8* dst, u8 src_size, u8 src2_size, u8 dst_size) {
+void text_gen3_concat(const u8* src, const u8* src2, u8* dst, size_t src_size, size_t src2_size, size_t dst_size) {
     text_general_concat(src, src2, dst, src_size, src2_size, dst_size, GEN3_EOL);
 }
 
-void text_gen2_concat(const u8* src, const u8* src2, u8* dst, u8 src_size, u8 src2_size, u8 dst_size) {
+void text_gen2_concat(const u8* src, const u8* src2, u8* dst, size_t src_size, size_t src2_size, size_t dst_size) {
     text_general_concat(src, src2, dst, src_size, src2_size, dst_size, GEN2_EOL);
 }
 
-void text_generic_terminator_fill(u8* src, u8 src_size) {
+void text_generic_terminator_fill(u8* src, size_t src_size) {
     text_general_terminator_fill(src, src_size, GENERIC_EOL);
 }
 
-void text_gen3_terminator_fill(u8* src, u8 src_size) {
+void text_gen3_terminator_fill(u8* src, size_t src_size) {
     text_general_terminator_fill(src, src_size, GEN3_EOL);
 }
 
-void text_gen2_terminator_fill(u8* src, u8 src_size) {
+void text_gen2_terminator_fill(u8* src, size_t src_size) {
     text_general_terminator_fill(src, src_size, GEN2_EOL);
 }
 
-void text_generic_replace(u8* src, u8 src_size, u8 base_char, u8 new_char) {
+void text_generic_replace(u8* src, size_t src_size, u8 base_char, u8 new_char) {
     text_general_replace(src, src_size, base_char, new_char, GENERIC_EOL);
 }
 
-void text_gen3_replace(u8* src, u8 src_size, u8 base_char, u8 new_char) {
+void text_gen3_replace(u8* src, size_t src_size, u8 base_char, u8 new_char) {
     text_general_replace(src, src_size, base_char, new_char, GEN3_EOL);
 }
 
-void text_gen2_replace(u8* src, u8 src_size, u8 base_char, u8 new_char) {
+void text_gen2_replace(u8* src, size_t src_size, u8 base_char, u8 new_char) {
     text_general_replace(src, src_size, base_char, new_char, GEN2_EOL);
 }
 
-void text_generic_copy(const u8* src, u8* dst, u8 src_size, u8 dst_size) {
+void text_generic_copy(const u8* src, u8* dst, size_t src_size, size_t dst_size) {
     text_general_copy(src, dst, src_size, dst_size, GENERIC_EOL);
 }
 
-void text_gen3_copy(const u8* src, u8* dst, u8 src_size, u8 dst_size) {
+void text_gen3_copy(const u8* src, u8* dst, size_t src_size, size_t dst_size) {
     text_general_copy(src, dst, src_size, dst_size, GEN3_EOL);
 }
 
-void text_gen2_copy(const u8* src, u8* dst, u8 src_size, u8 dst_size) {
+void text_gen2_copy(const u8* src, u8* dst, size_t src_size, size_t dst_size) {
     text_general_copy(src, dst, src_size, dst_size, GEN2_EOL);
 }
 
-u8 text_generic_is_same(const u8* src, const u8* dst, u8 src_size, u8 dst_size) {
+u8 text_generic_is_same(const u8* src, const u8* dst, size_t src_size, size_t dst_size) {
     return text_general_is_same(src, dst, src_size, dst_size, GENERIC_EOL);
 }
 
-u8 text_gen3_is_same(const u8* src, const u8* dst, u8 src_size, u8 dst_size) {
+u8 text_gen3_is_same(const u8* src, const u8* dst, size_t src_size, size_t dst_size) {
     return text_general_is_same(src, dst, src_size, dst_size, GEN3_EOL);
 }
 
-u8 text_gen2_is_same(const u8* src, const u8* dst, u8 src_size, u8 dst_size) {
+u8 text_gen2_is_same(const u8* src, const u8* dst, size_t src_size, size_t dst_size) {
     return text_general_is_same(src, dst, src_size, dst_size, GEN2_EOL);
 }
 
-u8 text_generic_count_question(const u8* src, u8 src_size) {
+size_t text_generic_count_question(const u8* src, size_t src_size) {
     return text_general_count_question(src, src_size, GENERIC_EOL, GENERIC_QUESTION);
 }
 
-u8 text_gen3_count_question(const u8* src, u8 src_size) {
+size_t text_gen3_count_question(const u8* src, size_t src_size) {
     return text_general_count_question(src, src_size, GEN3_EOL, GEN3_QUESTION);
 }
 
-u8 text_gen2_count_question(const u8* src, u8 src_size) {
+size_t text_gen2_count_question(const u8* src, size_t src_size) {
     return text_general_count_question(src, src_size, GEN2_EOL, GEN2_QUESTION);
 }
 
-u8 text_generic_size(const u8* src, u8 src_size) {
+size_t text_generic_size(const u8* src, size_t src_size) {
     return text_general_size(src, src_size, GENERIC_EOL);
 }
 
-u8 text_gen3_size(const u8* src, u8 src_size) {
+size_t text_gen3_size(const u8* src, size_t src_size) {
     return text_general_size(src, src_size, GEN3_EOL);
 }
 
-u8 text_gen2_size(const u8* src, u8 src_size) {
+size_t text_gen2_size(const u8* src, size_t src_size) {
     return text_general_size(src, src_size, GEN2_EOL);
 }
 
-void text_generic_to_gen3(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 UNUSED(jp_src), u8 jp_dst) {
+void text_generic_to_gen3(const u8* src, u8* dst, size_t src_size, size_t dst_size, u8 UNUSED(jp_src), u8 jp_dst) {
     if(jp_dst)
         text_general_conversion(src, dst, src_size, dst_size, GENERIC_EOL, GEN3_EOL, text_general_to_gen3_jp_bin);
     else
         text_general_conversion(src, dst, src_size, dst_size, GENERIC_EOL, GEN3_EOL, text_general_to_gen3_int_bin);
 }
 
-void text_gen3_to_generic(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 jp_src, u8 UNUSED(jp_dst)) {
+void text_gen3_to_generic(const u8* src, u8* dst, size_t src_size, size_t dst_size, u8 jp_src, u8 UNUSED(jp_dst)) {
     if(jp_src)
         text_general_conversion(src, dst, src_size, dst_size, GEN3_EOL, GENERIC_EOL, text_gen3_to_general_jp_bin);
     else
         text_general_conversion(src, dst, src_size, dst_size, GEN3_EOL, GENERIC_EOL, text_gen3_to_general_int_bin);
 }
 
-void text_gen3_to_gen12(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 jp_src, u8 jp_dst) {
+void text_gen3_to_gen12(const u8* src, u8* dst, size_t src_size, size_t dst_size, u8 jp_src, u8 jp_dst) {
     if(jp_src && jp_dst)
         text_general_conversion(src, dst, src_size, dst_size, GEN3_EOL, GEN2_EOL, text_gen3_to_gen12_jp_bin);
     else if(jp_src)
@@ -240,7 +241,7 @@ void text_gen3_to_gen12(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 jp_
         text_general_conversion(src, dst, src_size, dst_size, GEN3_EOL, GEN2_EOL, text_gen3_to_gen12_int_bin);
 }
 
-void text_gen12_to_gen3(const u8* src, u8* dst, u8 src_size, u8 dst_size, u8 jp_src, u8 UNUSED(jp_dst)) {
+void text_gen12_to_gen3(const u8* src, u8* dst, size_t src_size, size_t dst_size, u8 jp_src, u8 UNUSED(jp_dst)) {
     if(src[0] == GEN12_TRAINER)
         text_gen3_copy(get_table_pointer(trainer_names_bin, jp_src), dst, src_size, dst_size);
     else {
